Renamed RestUtils to RestNamingUtils
