+++ core/src/main/kotlin/org/evomaster/core/output/naming/rest/RestActionTestCaseNamingStrategy.kt
@@ -3,9 +3,9 @@ package org.evomaster.core.output.naming.rest
 import org.evomaster.core.output.naming.ActionTestCaseNamingStrategy
 import org.evomaster.core.output.naming.AmbiguitySolver
 import org.evomaster.core.output.naming.LanguageConventionFormatter
-import org.evomaster.core.output.naming.rest.RestUtils.addBodyShape
-import org.evomaster.core.output.naming.rest.RestUtils.getPath
-import org.evomaster.core.output.naming.rest.RestUtils.isGetCall
+import org.evomaster.core.output.naming.rest.RestNamingUtils.addBodyShape
+import org.evomaster.core.output.naming.rest.RestNamingUtils.getPath
+import org.evomaster.core.output.naming.rest.RestNamingUtils.isGetCall
 import org.evomaster.core.problem.rest.RestCallAction
 import org.evomaster.core.problem.rest.RestCallResult
 import org.evomaster.core.search.EvaluatedIndividual
